requirements.txt
fitz==0.0.1.dev2
langchain==0.2.1
python-dotenv==1.0.1
streamlit==1.34.0
